point_quality/requirements.txt
openai==2.7.1

httpx==0.28.1
anyio==4.11.0
sniffio==1.3.1
h11==0.16.0

pydantic==2.12.4
pydantic_core==2.41.5
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2

PyYAML==6.0.3
tqdm==4.67.1
